requirements.txt
torch==2.0.1
torchvision==0.15.2
timm==0.4.12
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
scipy
ftfy
opencv-python
sentencepiece
pyarrow
deepspeed==0.14.2
transformers
pytorch-lightning==2.2.1
nltk
rouge
scikit-image
accelerate
fairscale
ruamel.yaml
wandb
pandas
future
scikit-survival
torchmetrics==0.7.3
h5py
open_clip_torch>=0.2.1
pycocoevalcap
webdataset>=0.2.31
